pyproject.toml
[project]
name = "ai-interview-system"
version = "0.1.0"
description = "Graph-driven adaptive interviewing system for FMCG concept testing"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "POC Development Team"}
]

dependencies = [
    # Core
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "pydantic>=2.11.10,<=2.12.4",  # Compatible with Gradio 6.0.0
    "pydantic-settings>=2.6.0",

    # Graph
    "networkx>=3.2.1",

    # LLM Clients
    "anthropic>=0.40.0",
    "openai>=1.54.0",
    "httpx>=0.27.0",

    # UI
    "gradio>=6.0.0",

    # Utilities
    "python-dateutil>=2.8.2",

    # Optional: Visualization
    "matplotlib>=3.8.2",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "ipython>=8.18.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"
markers = [
    "integration: marks tests as integration tests (require real API keys)",
]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "respx>=0.21.0",
    "black>=23.12.1",
    "ruff>=0.1.11",
]
